+++ src/lib/index.ts
@@ -8,4 +8,7 @@ export { default as BasicLed } from './BasicLed'
 export { default as BasicList } from './BasicList'
 export { default as BasicEdit } from './BasicEdit'
 
+export { BasicDrawer, BasicDrawerHeader } from './BasicDrawer'
+export { default as BasicButton } from './BasicEdit'
+
 export { default as BasicAuth } from './BasicAuth'
